+++ main/src/api/types/mod.rs
@@ -3,7 +3,11 @@
 pub(crate) mod application_config_builder;
 pub(crate) mod backend;
 pub(crate) mod config;
+pub(crate) mod identity;
+pub(crate) mod ingress;
 pub(crate) mod loadbalancer_svc;
 pub(crate) mod outcome;
 pub(crate) mod pool;
+pub(crate) mod registry;
+pub(crate) mod scaling;
 pub(crate) mod strategy;