+++ main/src/core/pool/mod.rs
@@ -1,4 +1,5 @@
 //! Pool implementation layer.
 
 pub(crate) mod backend_pool_instance;
+pub(crate) mod handler_instance_pool;
 pub(crate) mod inner;