+++ main/src/core/registry/mod.rs
@@ -1,0 +1,4 @@
+//! Registry implementation layer.
+
+pub(crate) mod in_memory_pool_registry;
+pub(crate) mod toml_tenant_registry;